+++ common/src/main/java/com/wynntils/core/config/upfixers/impl/OverlayRestructuringUpfixer.java
@@ -15,6 +15,9 @@ public class OverlayRestructuringUpfixer extends RenamedPrefixesUpfixer {
                     "spellCastMessageOverlayFeature.spellCastMessageOverlay."),
             Pair.of("spellCastRenderFeature.", "spellCastVignetteFeature."),
             Pair.of("tokenTrackerFeature.tokenBarsOverlay.", "tokenBarsOverlayFeature.tokenBarsOverlay."),
+            Pair.of(
+                    "shamanTotemTrackingFeature.shamanTotemTimerOverlay.",
+                    "shamanTotemTimerOverlayFeature.shamanTotemTimerOverlay."),
             Pair.of("tokenTrackerFeature.", "tokenTrackerBellFeature."),
             Pair.of("arrowShieldTrackingFeature.", "arrowShieldTrackerOverlayFeature."),
             Pair.of("dustomPlayerListFeature.", "customPlayerListOverlayFeature."),
