+++ common/src/main/resources/assets/wynntils/lang/en_us.json
@@ -900,23 +900,25 @@
   "feature.wynntils.shamanMaskOverlay.overlay.shamanMask.maskDisplay.description": "How should the mask text be displayed? %mask% template will be replaced with your current mask.",
   "feature.wynntils.shamanMaskOverlay.overlay.shamanMask.maskDisplay.name": "Shaman Mask Display Template",
   "feature.wynntils.shamanMaskOverlay.overlay.shamanMask.name": "Shaman Mask Overlay",
+  "feature.wynntils.shamanTotemTimerOverlay.description": "Adds an overlay showing time left on your shaman totems.",
+  "feature.wynntils.shamanTotemTimerOverlay.name": "Shaman Totem Timer Overlay",
+  "feature.wynntils.shamanTotemTimerOverlay.overlay.shamanTotemTimer.firstTotemTextColor.description": "The color of the first totem timer",
+  "feature.wynntils.shamanTotemTimerOverlay.overlay.shamanTotemTimer.firstTotemTextColor.name": "First Totem Timer Color",
+  "feature.wynntils.shamanTotemTimerOverlay.overlay.shamanTotemTimer.name": "Shaman Totem Timer",
+  "feature.wynntils.shamanTotemTimerOverlay.overlay.shamanTotemTimer.secondTotemTextColor.description": "The color of the second totem timer",
+  "feature.wynntils.shamanTotemTimerOverlay.overlay.shamanTotemTimer.secondTotemTextColor.name": "Second Totem Timer Color",
+  "feature.wynntils.shamanTotemTimerOverlay.overlay.shamanTotemTimer.textShadow.description": "What should the text shadow look like?",
+  "feature.wynntils.shamanTotemTimerOverlay.overlay.shamanTotemTimer.textShadow.name": "Text Shadow",
+  "feature.wynntils.shamanTotemTimerOverlay.overlay.shamanTotemTimer.thirdTotemTextColor.description": "The color of the third totem timer",
+  "feature.wynntils.shamanTotemTimerOverlay.overlay.shamanTotemTimer.thirdTotemTextColor.name": "Third Totem Timer Color",
+  "feature.wynntils.shamanTotemTimerOverlay.overlay.shamanTotemTimer.totemTrackingDetail.description": "What extra totem detail should be displayed?",
+  "feature.wynntils.shamanTotemTimerOverlay.overlay.shamanTotemTimer.totemTrackingDetail.name": "Totem Tracking Detail",
   "feature.wynntils.shamanTotemTracking.description": "Adds the ability to track your shaman totems.",
   "feature.wynntils.shamanTotemTracking.firstTotemColor.description": "What color should the totem 1 highlight and tracking text be?",
   "feature.wynntils.shamanTotemTracking.firstTotemColor.name": "Shaman Totem 1 Color",
   "feature.wynntils.shamanTotemTracking.highlightShamanTotems.description": "Should your Shaman/Skyseer totem(s) be highlighted?",
   "feature.wynntils.shamanTotemTracking.highlightShamanTotems.name": "Highlight Shaman Totems",
   "feature.wynntils.shamanTotemTracking.name": "Shaman Totem Tracking",
-  "feature.wynntils.shamanTotemTracking.overlay.shamanTotemTimer.firstTotemTextColor.description": "The color of the first totem timer",
-  "feature.wynntils.shamanTotemTracking.overlay.shamanTotemTimer.firstTotemTextColor.name": "First Totem Timer Color",
-  "feature.wynntils.shamanTotemTracking.overlay.shamanTotemTimer.name": "Shaman Totem Timer",
-  "feature.wynntils.shamanTotemTracking.overlay.shamanTotemTimer.secondTotemTextColor.description": "The color of the second totem timer",
-  "feature.wynntils.shamanTotemTracking.overlay.shamanTotemTimer.secondTotemTextColor.name": "Second Totem Timer Color",
-  "feature.wynntils.shamanTotemTracking.overlay.shamanTotemTimer.textShadow.description": "What should the text shadow look like?",
-  "feature.wynntils.shamanTotemTracking.overlay.shamanTotemTimer.textShadow.name": "Text Shadow",
-  "feature.wynntils.shamanTotemTracking.overlay.shamanTotemTimer.thirdTotemTextColor.description": "The color of the third totem timer",
-  "feature.wynntils.shamanTotemTracking.overlay.shamanTotemTimer.thirdTotemTextColor.name": "Third Totem Timer Color",
-  "feature.wynntils.shamanTotemTracking.overlay.shamanTotemTimer.totemTrackingDetail.description": "What extra totem detail should be displayed?",
-  "feature.wynntils.shamanTotemTracking.overlay.shamanTotemTimer.totemTrackingDetail.name": "Totem Tracking Detail",
   "feature.wynntils.shamanTotemTracking.secondTotemColor.description": "What color should the totem 2 highlight and tracking text be?",
   "feature.wynntils.shamanTotemTracking.secondTotemColor.name": "Shaman Totem 2 Color",
   "feature.wynntils.shamanTotemTracking.thirdTotemColor.description": "What color should the totem 3 highlight and tracking text be?",
